+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* ExportFile can generate a geohash
+
 ## [1.0.1] - 2016-03-22
 ### Fixed
 * Rebuild